pyproject.toml
[project]
name = "mllam-data-interface"
version = "0.1.0"
description = "Specifications for machine learning weather datasets"
authors = [
    {name = "Leif Denby", email = "lcd@dmi.dk"},
]
dependencies = [
    "pyyaml>=6.0.1",
    "loguru>=0.7.2",
    "zarr>=2.17.2",
]
requires-python = ">=3.9"
readme = "README.md"
license = {text = "MIT"}

[build-system]
requires = ["pdm-backend"]
build-backend = "pdm.backend"


[tool.pdm]
distribution = true

[tool.pdm.dev-dependencies]
dev = [
    "pre-commit>=3.7.0",
]

[tool.isort]
profile = "black"

[tool.pdm.build]
source-includes = ["specs"]
